POCO:1.6.1-all: Fix 'warning C4389: '==' : signed/unsigned mismatch' from Util/src/HelpFormatter.cpp
Fix warning treated as error from HelpFormatter.cpp: 'warning C4389:
'==' : signed/unsigned mismatch'

This warning also reproed from latest released POCO1.7.6-all version and
this change should address it.

+++ Util/src/HelpFormatter.cpp
@@ -244,7 +244,7 @@ void HelpFormatter::formatText(std::ostream& ostr, const std::string& text, int 
 		}
 		else 
 		{
-			if (word.length() == maxWordLen)
+			if (word.length() == (size_t)maxWordLen)
 			{
 				clearWord(ostr, pos, word, indent);
 			}
